+++ apps/backend/src/libs/modules/plugins/auth/libs/helpers/check-is-white-route.helper.ts
@@ -1,6 +1,7 @@
 import { type HTTPMethod } from "~/libs/types/types.js";
 
 import { type WhiteRoute } from "../../auth.plugin.js";
+import { checkMatchUrlPattern } from "./check-match-url-pattern.helper.js";
 
 type CheckIsWhiteRouteParameters = {
 	method: HTTPMethod;
@@ -26,16 +27,7 @@ const checkIsWhiteRoute = ({
 			return false;
 		}
 
-		const rootPathCritea = "/*";
-		const isRootPath = path.includes(rootPathCritea);
-
-		if (!isRootPath) {
-			return url === path;
-		}
-
-		const rootPath = path.replace(rootPathCritea, "");
-
-		return url.startsWith(rootPath);
+		return checkMatchUrlPattern(path, url);
 	});
 };