+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -11,9 +11,11 @@ import { storage } from "~/libs/modules/storage/storage.js";
 import { toastNotifier } from "~/libs/modules/toast-notifier/toast-notifier.js";
 import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
 import {
-	pointsOfInterestApi,
+	pointOfInterestApi,
+	pointOfInterestReducer,
 	pointsOfInterestDetailsReducer,
 } from "~/modules/points-of-interest/points-of-interest.js";
+import { routeApi, reducer as routeReducer } from "~/modules/routes/routes.js";
 import { userApi, reducer as usersReducer } from "~/modules/users/users.js";
 
 import { type ExtraArguments, type RootReducer } from "./libs/types/types.js";
@@ -31,7 +33,8 @@ class Store {
 	public get extraArguments(): ExtraArguments {
 		return {
 			authApi,
-			pointsOfInterestApi,
+			pointOfInterestApi,
+			routeApi,
 			storage,
 			toastNotifier,
 			userApi,
@@ -50,7 +53,9 @@ class Store {
 			},
 			reducer: {
 				auth: authReducer,
+				pointsOfInterest: pointOfInterestReducer,
 				pointsOfInterestDetails: pointsOfInterestDetailsReducer,
+				route: routeReducer,
 				users: usersReducer,
 			},
 		});