+++ apps/frontend/src/modules/points-of-interest/slices/actions.ts
@@ -1,18 +1,35 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type APIResponse, type AsyncThunkConfig } from "~/libs/types/types.js";
-import { type PointsOfInterestResponseDto } from "~/modules/points-of-interest/points-of-interest.js";
+import {
+	type PointsOfInterestRequestDto,
+	type PointsOfInterestResponseDto,
+} from "~/modules/points-of-interest/points-of-interest.js";
 
-import { name as sliceName } from "./points-of-interest-details.slice.js";
+import { name as poiDetailssliceName } from "./points-of-interest-details.slice.js";
+import { name as poiSliceName } from "./points-of-interest.slice.js";
 
 const getById = createAsyncThunk<
 	APIResponse<PointsOfInterestResponseDto>,
 	number,
 	AsyncThunkConfig
->(`${sliceName}/load-by-id`, (id, { extra }) => {
-	const { pointsOfInterestApi } = extra;
+>(`${poiDetailssliceName}/load-by-id`, (id, { extra }) => {
+	const { pointOfInterestApi } = extra;
 
-	return pointsOfInterestApi.getById(id);
+	return pointOfInterestApi.getById(id);
 });
 
-export { getById };
+const create = createAsyncThunk<
+	APIResponse<PointsOfInterestResponseDto>,
+	PointsOfInterestRequestDto,
+	AsyncThunkConfig
+>(`${poiSliceName}/create`, async (payload, { extra }) => {
+	const { pointOfInterestApi, toastNotifier } = extra;
+
+	const pointOfInterest = await pointOfInterestApi.create(payload);
+	toastNotifier.showSuccess("Point of interest created successfully");
+
+	return pointOfInterest;
+});
+
+export { create, getById };