+++ packages/shared/src/modules/routes/libs/types/route-get-by-id-response-dto.type.ts
@@ -1,0 +1,11 @@
+type RouteGetByIdResponseDto = {
+	description: string;
+	id: number;
+	name: string;
+	pois: {
+		id: number;
+		visitOrder: number;
+	}[];
+};
+
+export { type RouteGetByIdResponseDto };