+++ tests/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [0.0.5](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/@smartscapes/tests-v0.0.4...@smartscapes/tests-v0.0.5) (2025-08-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @smartscapes/shared bumped from 1.19.0 to 1.20.0
+
+## [0.0.4](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/@smartscapes/tests-v0.0.3...@smartscapes/tests-v0.0.4) (2025-08-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @smartscapes/shared bumped from 1.18.0 to 1.19.0
+
+## [0.0.3](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/@smartscapes/tests-v0.0.2...@smartscapes/tests-v0.0.3) (2025-08-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @smartscapes/shared bumped from 1.17.0 to 1.18.0
+
 ## [0.0.2](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/@smartscapes/tests-v0.0.1...@smartscapes/tests-v0.0.2) (2025-08-12)